requirements.txt
upstash-redis
python-decouple
jupyter
replicate
pillow
fastapi
fastapi[standard]
# support .heic images -> see readme
# pillow-heif 